Don't require setting asset_host
Previously, croppable used asset_host in the image processing job to
know where to find the uploaded image blob. This requires a team to all
run the application on the same port in development, and might
complicate CDN deployments if assets aren't yet available on the CDN.

This passes the full root URL all the way down from the controller to
the job, which is a lot of handing off but seems workable.